requirements.txt
python-dotenv==1.0.1
langchain-ollama==0.2.3
PyMuPDF==1.23.26
langchain==0.1.12
langchain-core==0.1.32
langchain-community==0.0.27
langgraph==0.0.27
pyspellchecker==0.7.2
sentence-transformers==2.5.1
faiss-cpu==1.8.0
typing-extensions==4.9.0
pydantic==2.6.4
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
websockets==12.0